Renaming some components
