+++ src/common/Nav/SearchInput/index.js
@@ -10,7 +10,7 @@ import {
 } from "../../../features/search/searchSlice";
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
-import { fetchMovieById } from "../../../features/movies/moviePage/movieSlice";
+import { fetchMovieById } from "../../../features/movies/MovieTile/movieSlice";
 import { fetchPersonById } from "../../../features/people/PersonDetails/personSlice";
 import search from "../../../images/search.svg";
