+++ src/features/people/PeopleList/index.js
