+++ src/features/people/PeopleList/styled.js
