+++ src/features/people/PersonTile/styled.js
