+++ src/store.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "@redux-saga/core";
 import rootSaga from "./rootSaga";
 import popularMoviesReducer from "./features/movies/popularMoviesSlice";
-import movieReducer from "./features/movies/moviePage/movieSlice";
+import movieReducer from "./features/movies/MovieTile/movieSlice";
 import searchReducer from "./features/search/searchSlice";
 import popularPeopleReducer from "./features/people/popularPeopleSlice";
 import personReducer from "./features/people/PersonDetails/personSlice";
